refactor: unify host runtime ownership
