+++ apps/cli/src/init.test.ts
@@ -47,10 +47,12 @@ describe("initActor", () => {
     expect(source).toContain('const actorName = "reviewer"')
     expect(source).toContain("infer([")
     expect(server).toContain('import definition from "./actor"')
-    expect(server).toContain("layerActorThreads(definition")
-    expect(server).toContain("BunRuntime.runMain(Layer.launch(application))")
+    expect(server).toContain("const host = await createHost({")
+    expect(server).toContain("layersFor: () => layers")
+    expect(server).toContain("await host.close()")
+    expect(worker).toContain("createWorker(definition")
     expect(worker).toContain('import definition from "./actor"')
-    expect(worker).toContain('from "tardie/cloudflare"')
+    expect(worker).toContain('from "tardie/worker"')
     expect(worker).toContain('import { modelAdapters } from "tardie/model/adapter"')
     expect(worker).toContain('import { openAICompatibleAdapter } from "tardie/model/openai"')
     expect(worker).toContain("modelAdapters: modelAdapters(openAICompatibleAdapter)")
