+++ apps/server/src/api.ts
@@ -1,720 +1,1 @@
-import { Clock, Context, Duration, Effect, Layer, Schema, Stream } from "effect"
-import { HttpRouter, HttpServerRequest, HttpServerResponse } from "effect/unstable/http"
-import { HttpApiBuilder, type HttpApiEndpoint } from "effect/unstable/httpapi"
-import type { Event } from "@clavia/tardigrade-core/log/event"
-import type { ThreadEventRow } from "@clavia/tardigrade-core/log"
-import type { ActorThreadRecord } from "@clavia/tardigrade-core/actor"
-import { threadCreatedOf } from "@clavia/tardigrade-core/interaction/relations"
-import { invocationCoordinateOf } from "@clavia/tardigrade-core/interaction"
-import { methodRequestLocation, existingMethodRequest, prepareMethodRequest, methodRequestState, methodCancellationRequest, methodCancellationEvent } from "@clavia/tardigrade-host/transport/http/method-request"
-
-import {
-  Api,
-  apiOf,
-  InvalidRequest,
-  InvocationSettled,
-  invalidRequest,
-  ModelCatalogUnavailable,
-  RESERVED_ACTOR,
-  unacceptableField,
-  UnknownMethod,
-  UnknownMethodCall,
-  UnknownActor,
-  UnknownProjection,
-  UnknownThread,
-  type ActorSummary,
-  type ActorThread,
-  type ThreadAdded,
-  type ThreadsSnapshot,
-  type ProjectionDeclaration,
-  type ThreadNode
-} from "@clavia/tardigrade-client/contract"
-import { agentProjections } from "./actor"
-import { ModelCatalogStore } from "./catalog"
-import { providerAvailabilitiesOf } from "./catalog-availability"
-import { modelsPageOf, providersPageOf } from "./catalog-page"
-import { ServerConfig } from "./config"
-import { Threads, type ActorThreads } from "./host"
-import { publicThreadId, resolveThreadId } from "./thread-compat"
-import type { InferenceStream } from "./inference-stream"
-import { problemResponse } from "./problem"
-import { treeOf, type ThreadSummary } from "./projections"
-
-// The thread endpoints. A route is a lookup on the Threads service plus one projection, because the
-// read side is a pure function of a log (projections.ts) and the write side is one delivery
-// (host.ts). What each route accepts and answers is declared in contract.ts; this module is the
-// implementation of that declaration. Nothing here holds state between requests: the SSE tail keeps
-// a cursor for the connection it serves and nothing else, so two processes reading the same log
-// answer the same way.
-
-// The page size of GET /v1/threads/:id/events when the caller states no `limit`
-// (docs/how-to/server.md, "Endpoints").
-export const DEFAULT_EVENT_LIMIT = 200
-
-// How long an idle tail waits before writing a comment frame. A proxy between the client and this
-// process closes a connection that says nothing, and a comment is the cheapest thing to say.
-export const DEFAULT_SSE_HEARTBEAT = Duration.seconds(5)
-
-// DEFAULT_INFERENCE_STREAM_BUFFER_CAPACITY bounds unread transient frames per browser connection.
-export const DEFAULT_INFERENCE_STREAM_BUFFER_CAPACITY = 64
-
-export interface ApiOptions {
-  readonly limit?: number
-  readonly heartbeat?: Duration.Input
-  readonly inference?: InferenceStream
-  readonly inferenceBufferCapacity?: number
-}
-
-const paramOf = (params: Readonly<Record<string, string | undefined>>, name: string): string =>
-  params[name] ?? ""
-
-const singleOf = (value: string | ReadonlyArray<string> | undefined): string | undefined =>
-  value === undefined ? undefined : typeof value === "string" ? value : value[0]
-
-// A sequence number is a whole number at or above zero. The declared endpoints get this from their
-// query Schema (contract.ts, Seq); the stream is not a declared endpoint, so it reads its own.
-const integerOf = (raw: string | undefined): number | undefined => {
-  if (raw === undefined) return undefined
-  const trimmed = raw.trim()
-  return /^\d+$/.test(trimmed) ? Number(trimmed) : undefined
-}
-
-const unknownThreadDetail = (id: string) => `No thread named ${JSON.stringify(id)} has ever existed.`
-const unknownActorDetail = (id: string) => `No actor instance named ${JSON.stringify(id)} has ever existed.`
-
-const actorOf = (threads: Context.Service.Shape<typeof Threads>, id: string) =>
-  Effect.flatMap(threads.instance(id), (actor) =>
-    actor === undefined ? Effect.fail(UnknownActor.of(unknownActorDetail(id))) : Effect.succeed(actor))
-
-const unknownMethodDetail = (name: string, methods: Readonly<Record<string, unknown>>): string => {
-  const declared = Object.keys(methods)
-  const available = declared.length === 0
-    ? "This actor declares no methods."
-    : `This actor declares ${declared.map((method) => JSON.stringify(method)).join(", ")}.`
-  return `No method named ${JSON.stringify(name)} is declared. ${available}`
-}
-
-const failureMessage = (failure: unknown): string =>
-  failure instanceof Error ? failure.message : String(failure)
-
-// logOf reads a thread's events, failing the route when the log is empty. A thread exists once its
-// log has an event (docs/how-to/server.md, "Creation is delivery"), so an empty log is the only
-// unknown thread there is.
-const logOf = (read: (id: string) => Effect.Effect<ReadonlyArray<Event>>, id: string) =>
-  Effect.flatMap(read(id), (log) =>
-    log.length === 0 ? Effect.fail(UnknownThread.of(unknownThreadDetail(id))) : Effect.succeed(log))
-
-// flatten lists a forest depth-first, parent before child. The threads listing is this rather than
-// the raw thread list because `parent` is a fact of the forest and only treeOf can see it
-// (projections.ts, summaryOf).
-const flatten = (nodes: ReadonlyArray<ThreadNode>): ReadonlyArray<ThreadSummary> =>
-  nodes.flatMap(({ children, ...summary }) => [summary, ...flatten(children)])
-
-const findNode = (nodes: ReadonlyArray<ThreadNode>, id: string): ThreadNode | undefined => {
-  for (const node of nodes) {
-    if (node.id === id) return node
-    const found = findNode(node.children, id)
-    if (found !== undefined) return found
-  }
-  return undefined
-}
-
-const logsOf = (entries: ReadonlyArray<{ readonly id: string; readonly events: ReadonlyArray<Event> }>) =>
-  new Map(entries.map((entry) => [entry.id, entry.events] as const))
-
-const rosterOf = (threads: ActorThreads) =>
-  Effect.map(threads.list, (entries) => flatten(treeOf(logsOf(entries))))
-
-const frameOf = (seq: number, event: unknown): string => `id: ${seq}\ndata: ${JSON.stringify(event)}\n\n`
-
-// A comment frame: the client's parser drops it and the bytes keep the connection alive.
-const HEARTBEAT = ": tardigrade\n\n"
-
-const actorThreadOf = (record: ActorThreadRecord): ActorThread => ({
-  id: publicThreadId(record.thread),
-  ...(record.parentThread === undefined
-    ? {}
-    : { parent: publicThreadId(record.parentThread) }),
-  depth: record.depth
-})
-
-// openTails counts the SSE tails this process holds. A tail is a fiber that outlives its request
-// handler, so the count is what proves a disconnected client leaves nothing polling behind
-// (api.test.ts, "a reconnect replays from Last-Event-ID and then runs live, once each": the tail is
-// one while the client reads and zero once it aborts).
-let openTails = 0
-
-export const openStreams = (): number => openTails
-
-interface TailOptions {
-  readonly from: number
-  readonly limit: number
-  readonly heartbeat: Duration.Input
-  readonly initial?: Effect.Effect<readonly [string, number]>
-  readonly readPage: (cursor: number, limit: number) => Effect.Effect<ReadonlyArray<ThreadEventRow>>
-  readonly awaitHead: (cursor: number) => Effect.Effect<number>
-  readonly encodePage: (page: ReadonlyArray<ThreadEventRow>) => Effect.Effect<string>
-}
-
-const waitForHead = (
-  awaitHead: TailOptions["awaitHead"],
-  cursor: number,
-  heartbeat: Duration.Input
-) => Effect.race(
-  Effect.map(awaitHead(cursor), (target) => ({ kind: "commit" as const, target })),
-  Effect.as(Effect.sleep(heartbeat), { kind: "heartbeat" as const })
-)
-
-// resumableTail replays committed pages, waits at the head, and keeps an idle connection open.
-const resumableTail = (options: TailOptions): Stream.Stream<Uint8Array> => {
-  interface State {
-    readonly cursor: number
-    readonly waiting: boolean
-  }
-  const step = (state: State): Effect.Effect<readonly [string, State]> =>
-    Effect.gen(function*() {
-      let current = state
-      let target: number | undefined
-      for (;;) {
-        if (current.waiting) {
-          const wake = yield* waitForHead(options.awaitHead, current.cursor, options.heartbeat)
-          if (wake.kind === "heartbeat") return [HEARTBEAT, current] as const
-          target = wake.target
-          current = { ...current, waiting: false }
-        }
-        const page = yield* options.readPage(current.cursor, options.limit)
-        if (page.length > 0) {
-          const frames = yield* options.encodePage(page)
-          const cursor = page[page.length - 1]!.seq
-          return [frames === "" ? HEARTBEAT : frames, {
-            cursor,
-            waiting: target !== undefined && cursor >= target
-          }] as const
-        }
-        const wake = yield* waitForHead(options.awaitHead, current.cursor, options.heartbeat)
-        if (wake.kind === "heartbeat") return [HEARTBEAT, { ...current, waiting: true }] as const
-        target = wake.target
-      }
-    })
-  const framesFrom = (cursor: number) => Stream.unfold({ cursor, waiting: false } as State, step)
-  const frames = options.initial === undefined
-    ? framesFrom(options.from)
-    : Stream.unwrap(Effect.map(options.initial, ([frame, cursor]) =>
-      Stream.succeed(frame).pipe(Stream.concat(framesFrom(cursor)))))
-  return Stream.unwrap(
-    Effect.as(
-      Effect.acquireRelease(
-        Effect.sync(() => {
-          openTails += 1
-        }),
-        () =>
-          Effect.sync(() => {
-            openTails -= 1
-          })
-      ),
-      Stream.encodeText(frames)
-    )
-  )
-}
-
-const streamCursor = Effect.gen(function*() {
-  const query = yield* HttpServerRequest.ParsedSearchParams
-  const rawAfter = singleOf(query["after"])
-  const after = integerOf(rawAfter)
-  if (rawAfter !== undefined && after === undefined) {
-    return { problem: problemResponse(invalidRequest("Query", [unacceptableField("after")])) } as const
-  }
-  const request = yield* HttpServerRequest.HttpServerRequest
-  const rawLastEventId = request.headers["last-event-id"]
-  const lastEventId = integerOf(rawLastEventId)
-  if (rawLastEventId !== undefined && lastEventId === undefined) {
-    return { problem: problemResponse(invalidRequest("Headers", [unacceptableField("last-event-id")])) } as const
-  }
-  return { from: lastEventId ?? after } as const
-})
-
-const readableStreamOf = <A>(iterable: AsyncIterable<A>): ReadableStream<A> => {
-  const iterator = iterable[Symbol.asyncIterator]()
-  return new ReadableStream({
-    async pull(controller) {
-      const next = await iterator.next()
-      if (next.done) controller.close()
-      else controller.enqueue(next.value)
-    },
-    async cancel() {
-      await iterator.return?.()
-    }
-  })
-}
-
-const streamResponseOf = (body: Stream.Stream<Uint8Array>) => Effect.map(
-  Stream.toAsyncIterableEffect(body),
-  (iterable) => HttpServerResponse.raw(readableStreamOf(iterable), {
-    contentType: "text/event-stream",
-    headers: { "cache-control": "no-cache" }
-  })
-)
-
-// tail streams one thread with its durable sequence as both the page cursor and SSE id.
-const tail = (
-  readPage: ActorThreads["eventsPage"],
-  awaitHead: ActorThreads["awaitHead"],
-  id: string,
-  from: number,
-  limit: number,
-  heartbeat: Duration.Input
-): Stream.Stream<Uint8Array> => resumableTail({
-  from,
-  limit,
-  heartbeat,
-  readPage: (cursor, pageLimit) => readPage(id, cursor, pageLimit),
-  awaitHead: (cursor) => awaitHead(id, cursor),
-  encodePage: (page) => Effect.succeed(page.map(({ seq, event }) => frameOf(seq, event)).join(""))
-})
-
-const actorThreadsTail = (
-  threads: ActorThreads,
-  from: number | undefined,
-  limit: number,
-  heartbeat: Duration.Input
-): Stream.Stream<Uint8Array> => {
-  const initial = from === undefined
-    ? Effect.map(threads.actorThreads, ({ cursor, threads: records }) => [
-      frameOf(cursor, {
-        type: "ThreadsSnapshot",
-        threads: records.filter((record) => record.state === "registered").map(actorThreadOf)
-      } satisfies ThreadsSnapshot),
-      cursor
-    ] as const)
-    : undefined
-  return resumableTail({
-    from: from ?? 0,
-    limit,
-    heartbeat,
-    ...(initial === undefined ? {} : { initial }),
-    readPage: threads.actorEventsPage,
-    awaitHead: threads.awaitActorHead,
-    encodePage: (page) => Effect.map(
-      Effect.forEach(page, ({ seq, event }) => {
-        if (event.type !== "ThreadRegistered" || typeof event.thread !== "string") return Effect.succeed("")
-        return Effect.map(threads.actorThread(event.thread), (record) =>
-          record === undefined
-            ? ""
-            : frameOf(seq, { type: "ThreadAdded", thread: actorThreadOf(record) } satisfies ThreadAdded))
-      }),
-      (frames) => frames.join("")
-    )
-  })
-}
-
-// The stream stays an HttpRouter route rather than an HttpApi endpoint. HttpApi is
-// request-and-response shaped: an endpoint decodes a request, runs a handler, and encodes one
-// answer, while this route hands back a connection that outlives the handler and carries its own
-// cursor, heartbeat, and Last-Event-ID resume. It is merged beside the HttpApi app and inherits the
-// same bearer gate by being part of the same router (http.ts, layerApp).
-const streamResponse = (
-  threads: ActorThreads,
-  id: string,
-  limit: number,
-  heartbeat: Duration.Input
-) => Effect.gen(function*() {
-  const first = yield* threads.eventsPage(id, 0, 1)
-  if (first.length === 0) return problemResponse(UnknownThread.of(unknownThreadDetail(id)))
-  const cursor = yield* streamCursor
-  if ("problem" in cursor) return cursor.problem
-  return yield* streamResponseOf(tail(threads.eventsPage, threads.awaitHead, id, cursor.from ?? 0, limit, heartbeat))
-})
-
-const actorThreadsStreamResponse = (
-  threads: ActorThreads,
-  limit: number,
-  heartbeat: Duration.Input
-) => Effect.gen(function*() {
-  const cursor = yield* streamCursor
-  if ("problem" in cursor) return cursor.problem
-  return yield* streamResponseOf(actorThreadsTail(threads, cursor.from, limit, heartbeat))
-})
-
-const inferenceStreamResponse = (
-  inference: InferenceStream,
-  actor: string,
-  thread: string,
-  heartbeat: Duration.Input,
-  bufferCapacity: number
-) => Effect.sync(() => {
-  const encoder = new TextEncoder()
-  let unsubscribe: (() => void) | undefined
-  let heartbeatTimer: ReturnType<typeof setInterval> | undefined
-  const body = new ReadableStream<Uint8Array>({
-    start(controller) {
-      unsubscribe = inference.subscribe((delta) => {
-        if (delta.instance !== actor || delta.thread !== thread) return
-        if ((controller.desiredSize ?? 1) <= 0) return
-        controller.enqueue(encoder.encode(`data: ${JSON.stringify(delta)}\n\n`))
-      })
-      heartbeatTimer = setInterval(() => {
-        if ((controller.desiredSize ?? 1) > 0) controller.enqueue(encoder.encode(HEARTBEAT))
-      }, Duration.toMillis(heartbeat))
-    },
-    cancel() {
-      unsubscribe?.()
-      if (heartbeatTimer !== undefined) clearInterval(heartbeatTimer)
-    }
-  }, { highWaterMark: bufferCapacity })
-  return HttpServerResponse.raw(body, {
-    contentType: "text/event-stream",
-    headers: { "cache-control": "no-cache" }
-  })
-})
-
-export const layerStream = (options: ApiOptions = {}) => {
-  const limit = options.limit ?? DEFAULT_EVENT_LIMIT
-  const heartbeat = options.heartbeat ?? DEFAULT_SSE_HEARTBEAT
-  const bufferCapacity = options.inferenceBufferCapacity ?? DEFAULT_INFERENCE_STREAM_BUFFER_CAPACITY
-  if (!Number.isSafeInteger(bufferCapacity) || bufferCapacity <= 0) {
-    throw new Error(`inference stream buffer capacity must be a positive integer, got ${bufferCapacity}`)
-  }
-  return Layer.mergeAll(
-    HttpRouter.add(
-      "GET",
-      "/v1/actors/:id/threads/:thread/events/stream",
-      Effect.gen(function*() {
-        const params = yield* HttpRouter.params
-        const service = yield* Threads
-        const threads = yield* actorOf(service, paramOf(params, "id"))
-        return yield* streamResponse(threads, paramOf(params, "thread"), limit, heartbeat)
-      })
-    ),
-    HttpRouter.add(
-      "GET",
-      "/v1/actors/:id/threads/stream",
-      Effect.gen(function*() {
-        const params = yield* HttpRouter.params
-        const threads = yield* (yield* Threads).ensure(paramOf(params, "id"))
-        return yield* actorThreadsStreamResponse(threads, limit, heartbeat)
-      })
-    ),
-    ...(options.inference === undefined ? [] : [HttpRouter.add(
-      "GET",
-      "/v1/actors/:id/threads/:thread/inference/stream",
-      Effect.gen(function*() {
-        const params = yield* HttpRouter.params
-        const threads = yield* (yield* Threads).ensure(paramOf(params, "id"))
-        const thread = yield* resolveThreadId(paramOf(params, "thread"), (thread) => Effect.map(threads.actorThread(thread), (record) => record !== undefined))
-        return yield* inferenceStreamResponse(
-          options.inference!,
-          paramOf(params, "id"),
-          thread,
-          heartbeat,
-          bufferCapacity
-        )
-      })
-    )])
-  )
-}
-
-// layerThreadsGroup implements every declared thread endpoint over the Threads service and the
-// projections. It carries no gate of its own: the bearer middleware is global to the router, so a
-// route is inside it by being part of the same application (http.ts, layerAuth; http.test.ts, "a
-// token closes the API and leaves healthz open").
-export const layerThreadsGroup = (options: ApiOptions = {}) => {
-  const limit = options.limit ?? DEFAULT_EVENT_LIMIT
-  return HttpApiBuilder.group(Api, "threads", (handlers) =>
-    handlers
-      .handle("allocateRoot", ({ params, payload, query }) => Effect.gen(function* () {
-        const service = yield* Threads
-        if (query.actor !== undefined && query.actor !== (service.actorName ?? RESERVED_ACTOR)) return yield* Effect.fail(InvalidRequest.of("Allocation target actor does not match this deployment."))
-        const threads = yield* service.ensure(params.id)
-        return yield* threads.allocateRoot(payload.name, payload)
-      }))
-      // The body is the declared payload, decoded before this runs: a body that is not one is
-      // refused by the declaration and rendered as a problem document (contract.ts,
-      // layerRequestProblems), so the handler only ever sees an event.
-      .handle("append", ({ params, payload }) =>
-        Effect.gen(function*() {
-          const service = yield* Threads
-          const threads = yield* actorOf(service, params.id)
-          yield* logOf(threads.events, params.thread)
-          yield* threads.append(params.thread, payload)
-          return { actor: params.id, thread: params.thread }
-        }))
-      .handle("list", ({ params }) =>
-        Effect.gen(function*() {
-          const threads = yield* actorOf(yield* Threads, params.id)
-          return yield* rosterOf(threads)
-        }))
-      .handle("events", ({ params, query }) =>
-        Effect.gen(function*() {
-          const threads = yield* actorOf(yield* Threads, params.id)
-          const log = yield* logOf(threads.events, params.thread)
-          const { after, limit: page } = query
-          // The comma list is the one rule the query Schema does not state: every value it could
-          // hold is a valid event type, including ones this build has never seen.
-          const types = query.types?.split(",").map((type) => type.trim()).filter((type) => type.length > 0)
-          return log
-            .map((event, index) => ({ seq: index + 1, event }))
-            .filter((row) => row.seq > (after ?? 0) && (types === undefined || types.includes(row.event.type)))
-            .slice(0, page ?? limit)
-        }))
-      .handle("tree", ({ params }) =>
-        Effect.gen(function*() {
-          const threads = yield* actorOf(yield* Threads, params.id)
-          // The forest is built over every log because parentage is a claim in the PARENT's log; a
-          // subtree cannot be derived from the subtree's own events (projections.ts, treeOf).
-          const node = findNode(treeOf(logsOf(yield* threads.list)), params.thread)
-          if (node === undefined) {
-            return yield* Effect.fail(UnknownThread.of(unknownThreadDetail(params.thread)))
-          }
-          return node
-        })))
-}
-
-// layerRuntimeGroup describes the actor mounted at the runtime origin.
-export const layerRuntimeGroup = HttpApiBuilder.group(Api, "runtime", (handlers) =>
-  handlers.handle("metadata", () =>
-    Effect.map(Threads, (threads) => ({
-      name: threads.actorName ?? RESERVED_ACTOR,
-      storage: { kind: "sqlite", location: threads.sqlite }
-    }))))
-
-// ServerApi is the surface this process serves: the platform's log routes, actor methods, and declared projections.
-export const ServerApi = apiOf(agentProjections)
-
-// jsonSchemaOf attaches every generated definition to the root schema that references it.
-const jsonSchemaOf = (schema: Schema.Constraint): unknown => {
-  const document = Schema.toJsonSchemaDocument(schema)
-  return Object.keys(document.definitions).length === 0
-    ? document.schema
-    : { ...document.schema, $defs: document.definitions }
-}
-
-const methodOf = (threads: ActorThreads, name: string) => {
-  const method = threads.methods[name]
-  return method === undefined
-    ? Effect.fail(UnknownMethod.of(unknownMethodDetail(name, threads.methods)))
-    : Effect.succeed(method)
-}
-
-const invokeMethod = (
-  params: { readonly id: string; readonly thread: string; readonly method: string; readonly call: string },
-  query: { readonly actor?: string; readonly timeoutMs?: number },
-  payload: unknown
-) =>
-      Effect.gen(function*() {
-        const service = yield* Threads
-        if (query.actor !== undefined && query.actor !== (service.actorName ?? RESERVED_ACTOR)) return yield* Effect.fail(InvalidRequest.of("Invocation target actor does not match this deployment."))
-        const threads = yield* actorOf(service, params.id)
-        const method = yield* methodOf(threads, params.method)
-        const events = yield* logOf(threads.events, params.thread)
-        const reference = invocationCoordinateOf(
-          threadCreatedOf(events)?.address ?? { actor: service.actorName ?? RESERVED_ACTOR, instance: params.id, thread: params.thread },
-          { method: params.method, id: params.call, epoch: 0 }
-        )
-        const existing = existingMethodRequest(events, reference)
-        if (existing !== undefined) return existing
-        const at = yield* Clock.currentTimeMillis
-        const prepared = yield* Effect.try({
-          try: () => prepareMethodRequest({ reference, method, input: payload, at,
-            ...(query.timeoutMs === undefined ? {} : { timeoutMs: query.timeoutMs }) }),
-          catch: (failure) => InvalidRequest.of(failureMessage(failure))
-        })
-        yield* threads.append(params.thread, prepared.event)
-        return prepared.accepted
-      })
-
-// layerMethodsGroup invokes and reads the method declarations carried by the mounted actor runtime.
-export const layerMethodsGroup = HttpApiBuilder.group(ServerApi, "methods", (handlers) =>
-  handlers
-    .handle("methods", () =>
-      Effect.map(Threads, (threads) =>
-        Object.entries(threads.methods).map(([name, method]) => ({
-          name,
-          cancellable: method.cancellation !== undefined,
-          timeoutMs: method.timeoutMs,
-          inputSchema: jsonSchemaOf(method.input),
-          outputSchema: jsonSchemaOf(method.output)
-        }))))
-    .handle("invoke", ({ params, query, payload }) => invokeMethod(params, query, payload))
-    .handle("invokeMethod", ({ params, query, payload, headers }) => Effect.gen(function* () {
-      const call = headers["idempotency-key"]
-      if (!call.trim()) return yield* Effect.fail(InvalidRequest.of("Idempotency-Key must be a nonempty header"))
-      const receipt = yield* invokeMethod({ ...params, call }, query, payload)
-      return HttpServerResponse.jsonUnsafe(receipt, { status: 202, headers: { location: methodRequestLocation(receipt.reference) } })
-    }))
-    .handle("methodState", ({ params, query }) =>
-      Effect.gen(function*() {
-        const service = yield* Threads
-        if (query.actor !== undefined && query.actor !== (service.actorName ?? RESERVED_ACTOR)) {
-          return yield* Effect.fail(InvalidRequest.of("Invocation target actor does not match this deployment."))
-        }
-        const threads = yield* actorOf(service, params.id)
-        const method = yield* methodOf(threads, params.method)
-        const log = yield* logOf(threads.events, params.thread)
-        const { state } = methodRequestState(log, method, { method: params.method, id: params.call, ...(query.epoch === undefined ? {} : { epoch: query.epoch }) })
-        if (state === undefined) {
-          return yield* Effect.fail(
-            UnknownMethodCall.of(
-              `No call named ${JSON.stringify(params.call)} exists for method ${JSON.stringify(params.method)} on this thread.`
-            )
-          )
-        }
-        return state
-      }))
-    .handle("cancel", ({ params, query, payload }) =>
-      Effect.gen(function*() {
-        const service = yield* Threads
-        if (query.actor !== undefined && query.actor !== (service.actorName ?? RESERVED_ACTOR)) {
-          return yield* Effect.fail(InvalidRequest.of("Invocation target actor does not match this deployment."))
-        }
-        const threads = yield* actorOf(service, params.id)
-        const method = yield* methodOf(threads, params.method)
-        const log = yield* logOf(threads.events, params.thread)
-        const { invocation, status: disposition } = methodCancellationRequest(log, method, { method: params.method, id: params.call, ...(query.epoch === undefined ? {} : { epoch: query.epoch }) })
-        if (disposition === "unknown") {
-          return yield* Effect.fail(UnknownMethodCall.of(
-            `No call named ${JSON.stringify(params.call)} exists for method ${JSON.stringify(params.method)} on this thread.`
-          ))
-        }
-        if (disposition === "unsupported") {
-          return yield* Effect.fail(InvalidRequest.of(
-            `Method ${JSON.stringify(params.method)} does not declare cancellation.`
-          ))
-        }
-        if (disposition === "settled") {
-          return yield* Effect.fail(InvocationSettled.of(
-            `Invocation ${JSON.stringify(params.call)} has settled and cannot be cancelled.`
-          ))
-        }
-        if (disposition !== "requestable") {
-          return {
-            actor: params.id,
-            thread: params.thread,
-            method: params.method,
-            call: params.call,
-            status: disposition
-          }
-        }
-        const at = yield* Clock.currentTimeMillis
-        yield* threads.append(params.thread, methodCancellationEvent(invocation, at, payload.reason))
-        return {
-          actor: params.id,
-          thread: params.thread,
-          method: params.method,
-          call: params.call,
-          status: "requested" as const
-        }
-      })))
-
-// layerProjectionsGroup implements every declared projection the same way, because there is only
-// one way: read the thread's log, hand it to `run` with the decoded query, and answer what comes
-// back. Nothing about a projection's meaning reaches this module; the actor holds all of it.
-// readOf is the whole of what serving a projection is: refuse an actor this build does not serve,
-// read the thread's log, and hand it to `run` with the decoded query. Nothing about what a
-// projection means reaches this module; the actor holds all of it (actor.ts, agentProjections).
-const readOf = (declaration: ProjectionDeclaration) =>
-(request: {
-  readonly params: { readonly id: string; readonly thread: string }
-  readonly query: never
-}) =>
-  Effect.gen(function*() {
-    const threads = yield* actorOf(yield* Threads, request.params.id)
-    const log = yield* logOf(threads.events, request.params.thread)
-    return declaration.run(log, request.query)
-  })
-
-export const layerProjectionsGroup = HttpApiBuilder.group(ServerApi, "projections", (handlers) => {
-  // Every declared name gets the same handler, built from the same record the endpoints were
-  // generated from, so the two cannot disagree about which names exist. The type is the group's
-  // own endpoint map with every key required: `handleAll` accepts a partial record, and a partial
-  // one would leave the group short a handler at run time (api.test.ts, "a declared projection
-  // serves what the actor computes").
-  type Endpoints = (typeof handlers)["~EndpointsByIdentifier"]
-  type Complete = {
-    readonly [Name in keyof Endpoints]: HttpApiEndpoint.Handler<
-      Endpoints[Name],
-      HttpApiEndpoint.MiddlewareError<Endpoints[Name]>,
-      Threads
-    >
-  }
-  const served = Object.fromEntries(
-    Object.entries(agentProjections).map(([name, declaration]) => [name, readOf(declaration)])
-  ) as unknown as Complete
-  return handlers.handleAll(served)
-})
-
-// The name a request asked for when the actor never declared it.
-const declaredProjections = Object.keys(agentProjections)
-const declaredDetail = declaredProjections.length === 0
-  ? "This actor declares no projections."
-  : `This actor declares ${declaredProjections.map((name) => JSON.stringify(name)).join(", ")}.`
-
-export const layerUnknownProjection = HttpRouter.add(
-  "GET",
-  "/v1/actors/:id/threads/:thread/projections/:name",
-  Effect.gen(function*() {
-    const params = yield* HttpRouter.params
-    const name = paramOf(params, "name")
-    return problemResponse(
-      UnknownProjection.of(`No projection named ${JSON.stringify(name)} is mounted here. ${declaredDetail}`)
-    )
-  })
-)
-
-export const layerDefinitionsGroup = HttpApiBuilder.group(ServerApi, "definitions", (handlers) =>
-  handlers
-    .handle("definitions", () =>
-      Effect.flatMap(Threads, (threads): Effect.Effect<ReadonlyArray<ActorSummary>> =>
-        threads.definitions ?? Effect.succeed([{ name: RESERVED_ACTOR, builtIn: true }])))
-    .handle("pushDefinition", ({ payload }) =>
-      Effect.gen(function*() {
-        const threads = yield* Threads
-        if (threads.pushDefinition === undefined) {
-          return yield* Effect.fail(InvalidRequest.of("This server does not accept actor pushes."))
-        }
-        return yield* Effect.mapError(threads.pushDefinition(payload), (error) => InvalidRequest.of(error.message))
-      })))
-
-export const layerActorsGroup = HttpApiBuilder.group(ServerApi, "actors", (handlers) =>
-  handlers
-    .handle("actors", () => Effect.flatMap(Threads, (threads) => threads.instances))
-    .handle("ensureActor", ({ params }) =>
-      Effect.gen(function*() {
-        const threads = yield* Threads
-        yield* threads.ensure(params.id)
-        return { id: params.id, definition: threads.actorName ?? RESERVED_ACTOR }
-      }))
-    .handle("actor", ({ params }) =>
-      Effect.gen(function*() {
-        const threads = yield* Threads
-        yield* actorOf(threads, params.id)
-        return { id: params.id, definition: threads.actorName ?? RESERVED_ACTOR }
-      })))
-
-const catalogSnapshot = Effect.flatMap(ModelCatalogStore, (catalog) =>
-  catalog.snapshot === undefined
-    ? Effect.fail(ModelCatalogUnavailable.of(
-      "No validated model catalog is available. Check the server startup logs and catalog configuration."
-    ))
-    : Effect.succeed(catalog.snapshot))
-
-const catalogDiscovery = Effect.all([catalogSnapshot, ServerConfig]).pipe(
-  Effect.map(([catalog, config]) => ({
-    catalog,
-    availability: providerAvailabilitiesOf(config.model, config.modelCredentials),
-    policy: { ...(config.model.default === undefined ? {} : { default: config.model.default }), allow: config.model.allow }
-  }))
-)
-
-// layerModelsGroup pages the process snapshot using provider readiness derived without credential values.
-export const layerModelsGroup = HttpApiBuilder.group(ServerApi, "models", (handlers) =>
-  handlers
-    .handle("providers", ({ query }) => Effect.flatMap(catalogDiscovery, ({ catalog, availability, policy }) =>
-      Effect.try({
-        try: () => providersPageOf(catalog, availability, { ...query, policy }),
-        catch: (error) => InvalidRequest.of(failureMessage(error))
-      })))
-    .handle("models", ({ query }) => Effect.flatMap(catalogDiscovery, ({ catalog, availability, policy }) =>
-      Effect.try({
-        try: () => modelsPageOf(catalog, availability, { ...query, policy }),
-        catch: (error) => InvalidRequest.of(failureMessage(error))
-      }))))
+export * from "@clavia/tardigrade-http/api"
