+++ apps/web/src/docs/components/index.tsx
@@ -7,6 +7,7 @@ import { PeopleTalkingDiagram } from "./diagrams/PeopleTalkingDiagram"
 import { ActorCommunicationDiagram } from "./diagrams/ActorCommunicationDiagram"
 import { ActorInstancesDiagram } from "./diagrams/ActorInstancesDiagram"
 import { ChildThreadsDiagram } from "./diagrams/ChildThreadsDiagram"
+import { HostLayersDiagram } from "./diagrams/HostLayersDiagram"
 import { ThreadInvocationDiagram } from "./diagrams/ThreadInvocationDiagram"
 import { ThreadResolutionDiagram } from "./diagrams/ThreadResolutionDiagram"
 import { BehaviorTrajectoryDiagram } from "./diagrams/BehaviorTrajectoryDiagram"
@@ -248,6 +249,7 @@ export const mdxComponents = {
   PeopleTalkingDiagram,
   ChildThreadsDiagram,
   ThreadInvocationDiagram,
+  HostLayersDiagram,
   ThreadResolutionDiagram,
   BehaviorTrajectoryDiagram,
   Command,
