+++ examples/react-rlm-chat/server/tsconfig.json
@@ -5,7 +5,7 @@
     "resolveJsonModule": true,
     "types": ["@cloudflare/workers-types"],
     "paths": {
-      "tardie/cloudflare": ["../../../platform/cloudflare/src/index.ts"],
+      "tardie/worker": ["../../../platform/cloudflare/src/index.ts"],
       "tardie/model/*": ["../../../platform/model/src/*.ts"],
       "tardie/bun/*": ["../../../platform/bun/src/*.ts"],
       "tardie/server/*": ["../../../apps/server/src/*.ts"]
