+++ packages/tardie/src/bun.ts
@@ -1,0 +1,1 @@
+export * from "@clavia/tardigrade-bun"
