+++ platform/bun/package.json
@@ -34,7 +34,9 @@
     "@clavia/tardigrade-core": "workspace:*",
     "@clavia/tardigrade-host": "workspace:*",
     "@effect/sql-sqlite-bun": "4.0.0-rc.110",
-    "effect": "4.0.0-rc.110"
+    "effect": "4.0.0-rc.110",
+    "@clavia/tardigrade-http": "workspace:*",
+    "@effect/platform-bun": "4.0.0-rc.110"
   },
   "scripts": {
     "typecheck": "tsc --noEmit",
