+++ platform/bun/src/index.ts
@@ -1,3 +1,3 @@
-export { createHost, type Host, type HostOptions } from "./create-host"
+export { createHost, type Host, type HostOptions, type HostStorageLayout } from "./create-host"
 export { serve, DEFAULT_HOST_IDLE_TIMEOUT_SECONDS, DEFAULT_HOST_PORT, DEFAULT_HOST_HOSTNAME, type ServeOptions } from "./serve"
 export type { ClientThread, CallOptions } from "@clavia/tardigrade-core/actor/client"
