+++ tools/gate.ts
@@ -12,7 +12,7 @@ const root = fileURLToPath(new URL("../", import.meta.url))
 const pkg = (name: string) => `${root}packages/${name}`
 const packages = ["tardie", "core", "code", "agent", "host", "channels", "client"]
 const platformPkg = (name: string) => `${root}platform/${name}`
-const platforms = ["model", "bun", "worker-loader", "cloudflare"]
+const platforms = ["http", "model", "bun", "worker-loader", "cloudflare"]
 const appPkg = (name: string) => `${root}apps/${name}`
 const apps = ["cli", "server", "voyager"]
 const typecheckedApps = [...apps, "web"]
